Add ability to open format code generator in playlist view preferences
This adds ‘Format code generator’ items to the Tools menus on the Globals and Columns playlist view preferences tabs.

The items opens the same format code generator window that was previously available in Item details options.

The item shows only when a tab for a script where format codes can be used is active.
