+++ foo_ui_columns/foo_ui_columns.rc
@@ -757,9 +757,9 @@ BEGIN
     LTEXT           "Static",IDC_AXIS_RANGE,69,50,96,8
 END
 
-IDD_ITEM_DETAILS_PICK_FONT DIALOGEX 0, 0, 200, 227
+IDD_FORMAT_CODE_GENERATOR DIALOGEX 0, 0, 200, 227
 STYLE DS_SETFONT | DS_MODALFRAME | WS_POPUP | WS_VISIBLE | WS_CAPTION | WS_SYSMENU
-CAPTION "Format code generator – Item details"
+CAPTION "Format code generator"
 FONT 8, "Segoe UI", 400, 0, 0x0
 BEGIN
     LTEXT           "Font family",-1,14,7,47,8
@@ -1110,7 +1110,7 @@ BEGIN
         BOTTOMMARGIN, 95
     END
 
-    IDD_ITEM_DETAILS_PICK_FONT, DIALOG
+    IDD_FORMAT_CODE_GENERATOR, DIALOG
     BEGIN
         LEFTMARGIN, 14
         RIGHTMARGIN, 186
