+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -274,6 +274,7 @@
     <ClCompile Include="font_manager.cpp" />
     <ClCompile Include="font_manager_v3.cpp" />
     <ClCompile Include="font_picker.cpp" />
+    <ClCompile Include="format_code_generator.cpp" />
     <ClCompile Include="gdi.cpp" />
     <ClCompile Include="icons.cpp" />
     <ClCompile Include="imaging.cpp" />
@@ -461,6 +462,7 @@
     <ClInclude Include="file_info_utils.h" />
     <ClInclude Include="font_manager_data.h" />
     <ClInclude Include="font_picker.h" />
+    <ClInclude Include="format_code_generator.h" />
     <ClInclude Include="gdi.h" />
     <ClInclude Include="get_msg_hook.h" />
     <ClInclude Include="icons.h" />
