+++ foo_ui_columns/format_code_generator.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace cui::utils {
+
+HWND open_format_code_generator(HWND parent_wnd, GUID initial_font_id);
+
+}
