+++ foo_ui_columns/tab_global.cpp
@@ -1,6 +1,7 @@
 #include "pch.h"
 #include "ng_playlist/ng_playlist.h"
 #include "config.h"
+#include "format_code_generator.h"
 #include "help.h"
 
 extern const char* default_global_style_script;
@@ -113,6 +114,10 @@ public:
                     menu.append_separator();
                     menu.append_command(
                         collector.add([wnd] { help::open_text_styling_help(GetParent(wnd)); }), L"Text styling help");
+                    menu.append_command(collector.add([parent_wnd{GetAncestor(wnd, GA_ROOT)}] {
+                        utils::open_format_code_generator(parent_wnd, panels::playlist_view::items_font_id);
+                    }),
+                        L"Format code generator");
                 }
 
                 menu.append_separator();
